Save/Restore editor zoom level
The desire was to save the current zoom level and restore it when the application starts back up. To do this it first needed to ensure all editors are zooming in/out at the same time. Previously it was on a per editor bases. Zooming could be triggered two different ways, the first by the QActions which is easy to maintain and hook into to enforce all editors zooming in/out simultaneously and to save the zoom level. The second was within the ScintillaEditBase::wheelEvent() but that was not conducive to hook into.

The ZoomEventWatcher catches the specific wheel events before the ScintillaEditBase class can process it (it shares some code), but this way it can signal a zoom in/out event is happening. The MainWindow would then use this signal to translate it into to the aforementioned  QActions (which is easy to handle).

That then allowed for the zoom level to be stored, tracked, saved, loaded.

See #254