Embed translation files
Due to the various platform configurations (especially Linux) it makes it difficult to systematically locate the installed translation files. Since the number and size of the translation files are small, it makes more sense to embed these as qresources in the exe itself so loading the translation files can be deterministic.

Closes #155